moved logic to svgToPdf.ts; conversion from svg to pdf if renderMethod exists & pdf returned is not null
