+++ app/types/pdf.ts
@@ -1,0 +1,3 @@
+export type PDF = {
+  readonly filepath: string;
+}
